test: keep inactivity_slash_with_consecutive_epochs as its own suite
The two inactivity describes don't share a beforeAll (the InactivityTest
fixture is stateful per-instance), so merging them into one file only made
it longer without gaining shared setup. Split back into separate files and
revert the bootstrap.sh timeout bump that only applied to the merged file.
